fix: European slash format should not always be preferred
