+++ shared/editor/commands/table.ts
@@ -1,4 +1,3 @@
-import { parse, isValid } from "date-fns";
 import { GapCursor } from "prosemirror-gapcursor";
 import type { Node, NodeType } from "prosemirror-model";
 import { Slice } from "prosemirror-model";
@@ -20,6 +19,7 @@ import {
 } from "prosemirror-tables";
 import { ProsemirrorHelper } from "../../utils/ProsemirrorHelper";
 import { CSVHelper } from "../../utils/csv";
+import { parseDate } from "../../utils/date";
 import { chainTransactions } from "../lib/chainTransactions";
 import {
   getAllSelectedColumns,
@@ -82,10 +82,10 @@ export function createTableInner(
     const attrs =
       colWidth && index < colsCount - 1
         ? {
-          colwidth: [colWidth],
-          colspan: 1,
-          rowspan: 1,
-        }
+            colwidth: [colWidth],
+            colspan: 1,
+            rowspan: 1,
+          }
         : null;
     const cell = createCell(types.cell, attrs);
 
@@ -269,67 +269,6 @@ function isNullWidth({
   return !colwidth?.[0];
 }
 
-/**
- * Attempts to parse a date string in various common formats.
- *
- * @param dateStr The date string to parse.
- * @returns A Date object if parsing is successful, null otherwise.
- */
-function parseDate(dateStr: string): Date | null {
-  if (!dateStr) {
-    return null;
-  }
-
-  // Remove any trailing alphabetic text (e.g., "Uhr", "at", "o'clock", etc.)
-  const cleaned = dateStr.trim().replace(/\s*[a-zA-Z]+\s*$/g, "");
-
-  // Common date formats used in tables (with and without time, with and without year)
-  const formats = [
-    // ISO formats
-    "yyyy-MM-dd HH:mm:ss",
-    "yyyy-MM-dd HH:mm",
-    "yyyy-MM-dd",
-    // European dot formats
-    "dd.MM.yyyy HH:mm:ss",
-    "dd.MM.yyyy HH:mm",
-    "dd.MM.yyyy",
-    "dd.MM. HH:mm:ss",
-    "dd.MM. HH:mm",
-    "dd.MM.",
-    "d.M.yyyy HH:mm:ss",
-    "d.M.yyyy HH:mm",
-    "d.M.yyyy",
-    "d.M. HH:mm:ss",
-    "d.M. HH:mm",
-    "d.M.",
-    // European slash formats
-    "dd/MM/yyyy HH:mm:ss",
-    "dd/MM/yyyy HH:mm",
-    "dd/MM/yyyy",
-    "dd/MM HH:mm:ss",
-    "dd/MM HH:mm",
-    "dd/MM",
-    // US formats
-    "MM/dd/yyyy HH:mm:ss",
-    "MM/dd/yyyy HH:mm",
-    "MM/dd/yyyy",
-    "MM/dd HH:mm:ss",
-    "MM/dd HH:mm",
-    "MM/dd",
-  ];
-
-  const referenceDate = new Date();
-
-  for (const format of formats) {
-    const date = parse(cleaned, format, referenceDate);
-    if (isValid(date)) {
-      return date;
-    }
-  }
-
-  return null;
-}
-
 export function sortTable({
   index,
   direction,
@@ -375,14 +314,14 @@ export function sortTable({
 
       const nonEmptyCells = table
         .map((row) => row[index]?.textContent?.trim())
-        .filter((cell) => cell && cell.length > 0);
+        .filter((cell): cell is string => !!cell && cell.length > 0);
       if (nonEmptyCells.length > 0) {
         // check if all non-empty cells are valid dates
-        compareAsDate = nonEmptyCells.every((cell) => parseDate(cell!) !== null);
+        compareAsDate = nonEmptyCells.every((cell) => parseDate(cell) !== null);
         // if not dates, check if all non-empty cells are numbers
         if (!compareAsDate) {
           compareAsNumber = nonEmptyCells.every(
-            (cell) => !isNaN(parseFloat(cell!))
+            (cell) => !isNaN(parseFloat(cell))
           );
         }
       }
@@ -393,8 +332,12 @@ export function sortTable({
         const bContent = b[index]?.textContent ?? "";
 
         // empty cells always go to the end
-        if (!aContent) { return bContent ? 1 : 0; }
-        if (!bContent) { return -1; }
+        if (!aContent) {
+          return bContent ? 1 : 0;
+        }
+        if (!bContent) {
+          return -1;
+        }
 
         if (compareAsDate) {
           const aDate = parseDate(aContent);
